Implement global and local python functions for optimization
* optimization functions (cost and constraints) can
  be formulated by set of python functions.
* for global and local functions
* see test cases python_function and local_python_function

+++ source/Domain/CoefFunction/CoefFunctionFileData.cc
@@ -124,7 +124,7 @@ void CoefFunctionFileData::ReadData(std::istream& input, int dim)
   }
 
 
-  assert(idx == node_.GetSize());
+  assert(idx == (int) node_.GetSize());
 }
